separate buildTarget
